requirements.txt
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
flake8==4.0.1
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
jeepney==0.8.0
jira==3.2.0
jsonpickle==2.2.0
keyring==23.6.0
kiwisolver==1.4.3
matplotlib==3.5.2
mccabe==0.6.1
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
